backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
boto3=1.34.127=pypi_0
botocore=1.34.127=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.3.11=hca03da5_0
certifi=2024.2.2=pypi_0
click=8.1.7=pypi_0
dnspython=2.6.1=pypi_0
email-validator=2.1.1=pypi_0
expat=2.6.2=h313beb8_0
fastapi=0.111.0=pypi_0
fastapi-cli=0.0.4=pypi_0
fpdf=1.7.2=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
idna=3.7=pypi_0
jinja2=3.1.4=pypi_0
jmespath=1.0.1=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
llvmlite=0.42.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
mdurl=0.1.2=pypi_0
ncurses=6.4=h313beb8_0
numba=0.59.1=pypi_0
numpy=1.26.4=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.0.13=h1a28f6b_2
orjson=3.10.3=pypi_0
pdf2image=1.17.0=pypi_0
pillow=10.3.0=pypi_0
pip=24.0=py312hca03da5_0
pydantic=2.7.2=pypi_0
pydantic-core=2.18.3=pypi_0
pygments=2.18.0=pypi_0
pypdf2=3.0.1=pypi_0
python=3.12.3=h99e199e_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.9=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h1a28f6b_0
rich=13.7.1=pypi_0
s3transfer=0.10.1=pypi_0
setuptools=69.5.1=py312hca03da5_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h80987f9_0
starlette=0.37.2=pypi_0
tk=8.6.14=h6ba3021_0
typer=0.12.3=pypi_0
typing-extensions=4.12.1=pypi_0
tzdata=2024a=h04d1e81_0
ujson=5.10.0=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.30.0=pypi_0
uvloop=0.19.0=pypi_0
watchfiles=0.22.0=pypi_0
websockets=12.0=pypi_0
wheel=0.43.0=py312hca03da5_0
xz=5.4.6=h80987f9_1
zlib=1.2.13=h18a0788_1
